Claim Overtime, Overtime Claims Submission list, Overtime Claims Approval

+++ src/main/java/com/team7/leave/model/Employee.java
@@ -30,7 +30,7 @@ public class Employee {
 	private String password;
 	private Integer leaveMedicalLeft;
 	private Integer leaveAnnualLeft;
-	private Integer otHours;
+	private Double otHours;
 	@Column(name = "managedby")
 	private String managedBy;
 	
@@ -45,7 +45,7 @@ public class Employee {
 	
 
 	public Employee(String name, String email, String username, String password, Integer leaveMedicalLeft,
-			Integer leaveAnnualLeft, Integer otHours, String managedBy, Set<LeaveApplication> leaveList,
+			Integer leaveAnnualLeft, Double otHours, String managedBy, Set<LeaveApplication> leaveList,
 			Set<Overtime> overtimeList, UserType usertype) {
 		super();
 		this.name = name;
@@ -138,12 +138,12 @@ public class Employee {
 	}
 
 
-	public Integer getOtHours() {
+	public Double getOtHours() {
 		return otHours;
 	}
 
 
-	public void setOtHours(Integer otHours) {
+	public void setOtHours(Double otHours) {
 		this.otHours = otHours;
 	}
